+++ doubleclick-openrtb/src/main/java/com/google/doubleclick/openrtb/DoubleClickOpenRtbMapper.java
@@ -22,7 +22,6 @@ import static java.lang.Math.min;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
-import com.google.common.io.BaseEncoding;
 import com.google.doubleclick.DcExt;
 import com.google.doubleclick.util.CityDMARegionKey;
 import com.google.doubleclick.util.CityDMARegionValue;
@@ -59,7 +58,6 @@ import com.google.openrtb.OpenRtb.ContentCategory;
 import com.google.openrtb.json.OpenRtbJsonFactory;
 import com.google.openrtb.mapper.OpenRtbMapper;
 import com.google.openrtb.util.OpenRtbUtils;
-import com.google.protobuf.ByteString;
 import com.google.protobuf.TextFormat;
 import com.google.protos.adx.NetworkBid;
 
@@ -71,7 +69,6 @@ import org.slf4j.LoggerFactory;
 
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.Base64;
 import java.util.Calendar;
 import java.util.EnumSet;
 import java.util.LinkedHashSet;
@@ -137,7 +134,7 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
 
   @Override public OpenRtb.BidRequest.Builder toOpenRtbBidRequest(NetworkBid.BidRequest dcRequest) {
     OpenRtb.BidRequest.Builder request = OpenRtb.BidRequest.newBuilder()
-        .setId(Base64.getEncoder().encodeToString(dcRequest.getId().toByteArray()));
+        .setId(MapperUtil.toBase64(dcRequest.getId()));
 
     if (dcRequest.getIsPing()) {
       return request;
@@ -228,18 +225,14 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
       NetworkBid.BidRequest.Mobile dcMobile = dcRequest.getMobile();
 
       if (dcMobile.hasAdvertisingId()) {
-        device.setIfa(BaseEncoding.base64().encode(
-            dcMobile.getAdvertisingId().toByteArray()));
+        device.setIfa(MapperUtil.toBase64(dcMobile.getAdvertisingId()));
       } else if (dcMobile.hasEncryptedAdvertisingId()) {
-        device.setIfa(BaseEncoding.base64().encode(
-            dcMobile.getEncryptedAdvertisingId().toByteArray()));
+        device.setIfa(MapperUtil.toBase64(dcMobile.getEncryptedAdvertisingId()));
       }
       if (dcMobile.hasHashedIdfa()) {
-        device.setDpidmd5(BaseEncoding.base64().encode(
-            dcMobile.getHashedIdfa().toByteArray()));
+        device.setDpidmd5(MapperUtil.toBase64(dcMobile.getHashedIdfa()));
       } else if (dcMobile.hasEncryptedHashedIdfa()) {
-        device.setDpidmd5(BaseEncoding.base64().encode(
-            dcMobile.getEncryptedHashedIdfa().toByteArray()));
+        device.setDpidmd5(MapperUtil.toBase64(dcMobile.getEncryptedHashedIdfa()));
       }
     }
 
@@ -817,10 +810,9 @@ public class DoubleClickOpenRtbMapper implements OpenRtbMapper<
 
     if ((coppa && dcRequest.hasConstrainedUsageHostedMatchData())
         || (!coppa && dcRequest.hasHostedMatchData())) {
-      ByteString dcHMD = coppa
+      user.setCustomdata(MapperUtil.toBase64(coppa
           ? dcRequest.getConstrainedUsageHostedMatchData()
-          : dcRequest.getHostedMatchData();
-      user.setCustomdata(Base64.getEncoder().encodeToString(dcHMD.toByteArray()));
+          : dcRequest.getHostedMatchData()));
     }
 
     if (dcRequest.hasUserDemographic()) {